Rock, Paper, Scissors!-added-game winner loop break
added conditionals that stop the for loop when the player or computer score === 3 
this conditional will also log (GAME OVER!) 
added alert(PLAYER WINs! or COMPUTER WINS!) depending on which player reaches  a score of 3 first.
console logs (Game Over!) and final score displayed then insert a (break;) to exit the loop and prevent playRound() from executing and showing prompt.